Move validators into separate directory
